Add shows page back.
