+++ src/components/Info.vue
@@ -14,25 +14,8 @@
 				<span class="zip" itemprop="postalCode">{{ info.zip }}</span>
 			</p>
 			<p><a class="bandcamp img" v-bind:href="info.bandcampURL"><span class="icon"></span> <span class="text">bandcamp</span></a></p>
+			<p class="small">this site is <a href="https://github.com/kangabell/mar">open source</a>.</p>
 
-		</section>
-
-		<section>
-
-			<h2>Past Shows</h2>
-
-			<p v-for="show in reverseChronShows" v-if="show.archive !== false">
-				<strong>{{ show.date | formatDate }}</strong> &#151; 
-				<span v-if="show.note">{{ show.note }}</span>
-				<span v-else>
-					<span v-if="show.bands" v-for="(band, index) in show.bands">{{ band.name }}<span v-if="index+1 < show.bands.length">, </span></span> <!-- comma-separated list of bands -->
-				</span>
-				@ {{ show.location }}
-			</p>
-		</section>
-
-		<section>
-			<p class="small">this site is <a href="https://github.com/kangabell/mar">open source</a>, made by kay.</p>
 		</section>
 
 	</main>
@@ -43,9 +26,6 @@
 <script>
 
 	import {db} from '../firebase';
-	import _ from 'lodash'
-	import moment from 'moment'
-	
 
 	export default {
 		data() {
@@ -54,9 +34,7 @@
 				articles: {},
 				members: {},
 				vendors: {},
-				info: {},
-				announcements: {},
-				shows: {}
+				info: {}
 			}
 		},
 		firebase: {
@@ -75,28 +53,6 @@
 			info: {
 				source: db.ref('info'),
 				asObject: true
-			},
-			announcements: {
-				source: db.ref('announcements'),
-			},
-			shows: {
-				source: db.ref('shows'),
-			}
-		},
-		computed: {
-		  chronShows: function () {
-		    return _.orderBy(this.shows, 'date')
-		  },
-		  reverseChronShows: function () {
-		    return _.orderBy(this.shows, 'date').reverse();
-		  }
-		},
-		filters: {
-			formatDate: function(value) {
-				if (value) return moment(String(value)).format('MMMM D YYYY')
-			},
-			formatDateShort: function(value) {
-				if (value) return moment(String(value)).format('MMMM D')
 			}
 		}
 	}
