+++ src/components/header.vue
@@ -4,6 +4,7 @@
 
 		<nav>
 			<router-link to="/" exact>Home</router-link>
+			<router-link to="/shows" exact>Shows</router-link>
 			<router-link to="/music" exact>Music</router-link>
 			<router-link to="/merch" exact>Merch</router-link>
 			<router-link to="/info" exact>Info</router-link>
